+++ matsim/src/main/java/org/matsim/core/network/NetworkUtils.java
@@ -42,6 +42,7 @@ import org.matsim.core.gbl.Gbl;
 import org.matsim.core.network.algorithms.NetworkModeRestriction;
 import org.matsim.core.network.algorithms.NetworkSimplifier;
 import org.matsim.core.network.io.MatsimNetworkReader;
+import org.matsim.core.network.turnRestrictions.DisallowedNextLinks;
 import org.matsim.core.utils.geometry.CoordUtils;
 import org.matsim.core.utils.geometry.CoordinateTransformation;
 import org.matsim.core.utils.misc.OptionalTime;
