+++ matsim/src/main/java/org/matsim/core/network/algorithms/NetworkMergeDoubleLinks.java
@@ -28,8 +28,8 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.network.Node;
 import org.matsim.core.api.internal.NetworkRunnable;
-import org.matsim.core.network.DisallowedNextLinks;
-import org.matsim.core.network.DisallowedNextLinksUtils;
+import org.matsim.core.network.turnRestrictions.DisallowedNextLinks;
+import org.matsim.core.network.turnRestrictions.DisallowedNextLinksUtils;
 import org.matsim.core.network.NetworkUtils;
 
 import java.util.*;
