+++ matsim/src/test/java/org/matsim/core/network/DisallowedNextLinksTest.java
@@ -16,6 +16,7 @@ import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.network.Node;
 import org.matsim.core.network.io.MatsimNetworkReader;
 import org.matsim.core.network.io.NetworkWriter;
+import org.matsim.core.network.turnRestrictions.DisallowedNextLinks;
 import org.matsim.utils.objectattributes.attributeconverters.DisallowedNextLinksAttributeConverter;
 
 public class DisallowedNextLinksTest {
