+++ matsim/src/test/java/org/matsim/core/network/DisallowedNextLinksUtilsTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
+import org.matsim.core.network.turnRestrictions.DisallowedNextLinks;
+import org.matsim.core.network.turnRestrictions.DisallowedNextLinksUtils;
 
 public class DisallowedNextLinksUtilsTest {
